backend/requirements.txt
fastapi
uvicorn
pydantic
openai 
tiktoken 
chromadb 
langchain 
pypdf 
rapidocr-onnxruntime
requests 
PyPDF2 
pandas   
python-dotenv   
aiohttp


